addon(gameinput): report keyboard layouts as unsigned KLIDs
IME and custom layouts such as 0xE0010409 set the top bit of the
32-bit KLID, and get_keyboard_layout() returned it through a signed
int, so they came back negative. The accessor and the test seam now
use int64. The layout in get_device_info() is the tracked one, since
GetDeviceInfo() is static and goes stale when the player switches
input language.

Copilot-Session: 3e648c53-d612-4ce3-9f3e-46f8b4df51b3

+++ addons/godot_gameinput/src/gameinput_device.cpp
@@ -376,7 +376,7 @@ int GameInputDevice::get_system_buttons() const {
     return gi ? gi->device_get_system_buttons(m_id) : (int)SYSTEM_BUTTON_NONE;
 }
 
-int GameInputDevice::get_keyboard_layout() const {
+int64_t GameInputDevice::get_keyboard_layout() const {
     GameInput *gi = GameInput::get_singleton();
     return gi ? gi->device_get_keyboard_layout(m_id) : 0;
 }
